@@ -1,10 +1,15 @@
 module github.com/cybozu-go/netutil
 
+go 1.19
+
 require (
-	github.com/google/go-cmp v0.5.8
-	github.com/onsi/gomega v1.20.0
-	github.com/vishvananda/netlink v1.1.0
-	golang.org/x/net v0.0.0-20220822230855-b0a4917ee28c
+	github.com/google/go-cmp v0.5.9
+	github.com/onsi/gomega v1.25.0
+	github.com/vishvananda/netlink v1.2.1-beta.2
+	golang.org/x/net v0.5.0
 )
 
-go 1.16
+require (
+	github.com/vishvananda/netns v0.0.3 // indirect
+	golang.org/x/sys v0.4.0 // indirect
+)
